WIP: Splitting development build into ssr and web to reflect the use case
