+++ pages/get-text-content-page.ts
