+++ pages/get-text-sample.function.ts
@@ -1,9 +1,9 @@
 import { LoaderBuilder } from '../src/loaders';
 
 export default async function getTextSample(): Promise<string> {
-  const url = '../text-samples/ping.txt';
+  const url = './text-samples/ping.txt';
 
-  const builder = await LoaderBuilder.fromWebHost(url);
+  const builder = await LoaderBuilder.fromFile(url);
 
   if (url.includes('.md')) {
     await builder.asMarkdown();
