+++ pages/pages-json-location.const.ts
@@ -1,0 +1,1 @@
+export const pagesJsonLocation = './pages.json';
