+++ scripts/start-dev-server.sh
@@ -1,0 +1,8 @@
+rm -rf dist
+webpack --config-name development-web development-server
+
+chmod +x ./dist/buildPages.js
+
+./dist/buildPages.js
+
+http-server -c-1 ./dist
