+++ tsconfig.json
@@ -2,8 +2,8 @@
   "compilerOptions": {
     "esModuleInterop": true,
     "outDir": "dist",
-    "target": "ES2020",
-    "module": "ES2020",
+    "target": "ES2022",
+    "module": "ES2022",
     "moduleResolution": "node",
     "plugins": [
       {
