+++ webpack/prod/index.js
@@ -1,0 +1,4 @@
+module.exports = [
+  require('./webpack.config.server'),
+  require('./webpack.config.web'),
+];
