+++ webpack/prod/webpack.config.web.js
@@ -1,0 +1,53 @@
+const CopyWebpackPlugin = require('copy-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+
+module.exports = {
+  name: 'production-web',
+  mode: 'production',
+  entry: {
+    components: './src/components/index.ts',
+  },
+  output: {
+    filename: '[name].js',
+    library: {
+      type: 'commonjs-module',
+    },
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'ts-loader',
+          options: {
+            configFile: 'tsconfig.prod.json',
+          },
+        },
+      },
+      {
+        test: /\.css$/i,
+        use: [
+          'style-loader',
+          {
+            loader: 'css-loader',
+            options: {
+              importLoaders: 1,
+              modules: true,
+            },
+          },
+        ],
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.ts', '.js'],
+  },
+  plugins: [
+    new CopyWebpackPlugin({
+      patterns: ['package.json'],
+    }),
+    new MiniCssExtractPlugin(),
+  ],
+  target: 'web',
+};
